*sigh*
These tabs make things SO much harder (for people editing the generated
Makefile), and well, everything should ideally be on one line anyways.
Really reallly really a BIG pet peeve (somewhat like a large gerbil, if it
were in your boxers, munching at the elastic).

But that's not the actual problem at hand.  Linking libkdecore with
-lkdecore is bad... mmm'kay?

This means that:

a.) it will link in a possibly old and binary incompatible kdecore
b.) it will fail if you haven't installed kdelibs already

The proper solution:

$(LIB_KDECORE)

svn path=/trunk/kdelibs/; revision=63120