Implemented the "show SSL information" button in the dialog box when entering SSL mode.
Some infrastructure work was needed for this :
* Registering jobs to uiserver on demand of the slave (signal needProgressId)
* Storing the progress id in the slave
* Letting the uiserver access the metadata of a job
* Made SlaveBase::sendMetaData protected so that the slave can force the sending
(Waldo, I hope this is ok ?)
* Adding type SSL to messageBox()

I made the ssl info dlg modal so that the page doesn't load until one clicks
Close, but maybe we should then pop up the msg box again, and have a Cancel
button (in case we see a not-so-good certificate ?). Georges ? WDYT ?

PS: Restart KDE after compiling this.

svn path=/trunk/kdelibs/; revision=65974