Show queries of an mailto URL in a nice way when the mouse is over an mailto URL (for those of us who don't feel like learning how to read a URL); first propose at how to support the HREFLANG attribute of the A element.
The HREFLANG attribute holds one of the known two-letter country codes (de,
us, cn, ...) and thereby gives information in what language the page is
written which the HREF attribute points to.

My idea was it to show the correct flag for the HREFLANG in front of the URL
in the statusbar whenever the mouse hovers above an http URL which features
the HREFLAN attribute, e.g. when the mouse hovers above

<a href="http://www.kde.de" hreflang="de">http://www.kde.de</a>

you should see a german flag followed by the URL in the status bar.
Currently, everything seems to work fine, except that no flag his shown but
(for this case) [de] instead.

svn path=/trunk/kdelibs/; revision=68824