Okay let see here...
- Changed the way I was going to keep ref count when application cached
  password.  Now this has to be explicitly registered through the newly
  added method registerCachedAuthKey.  The previous code to do this was
  completely wrong.

- Fixed a problem when one person logins into the same site with two
  different usernames.  Now the last one will always override the first
  since the key used to cache them is the same.  Hmmm... will have to
  test this for caching that is not based on "Realm" value, ex ftp.

TODO: Figure out how to remove the cached passwords once the applications
using them are closed.  We do not really want to force something like this
on app developers.  David suggested a while back that I check whether
this can be added to the io-slave scheduler since every app that uses the
io-slave only has one copy of it. Problem is how to catch the metadata that
is sent whenever a new authorization info is cached from the scheduler ?

svn path=/trunk/kdelibs/; revision=69027