- Improved support for the hreflang attribute - now, a nice flag is shown in the statusbar instead of the country code. - Show a "send-mail" image in the statusbar for mailto urls, to make it easier to distinguish between mailto and other urls. - Fixed a typo in khtml_part.h (for the docu, that is)
Sidenote: the current way of drawing pixmaps in the statusbar (by making use
of Qt's richtext support) is rather buggy as the pixmaps aren't centered
vertically in the statusbar, but the whole richtext (As a unit, pixmap + raw
text) is. OTOH, it seems overkill to me to add a whole new method to KParts
just for this tiny (and, I admit, hardly used feature).

svn path=/trunk/kdelibs/; revision=69237