void setAutoSaveSettings( const QString & groupName, bool saveWindowSize );
This enables the "auto save *bar settings [and window size] when closing
the window, if the settings changed", as discussed with Waldo.
The bool for window size takes care of the difference between konqueror
(has another way of saving it) and e.g. KMail (which will want to use
this to get faster closing of windows)

Many thanks to Reggie for the signal in QMainWindow called endMovingToolBar :)

svn path=/trunk/kdelibs/; revision=70792