Implementation of attribute change notifications. It works like
  connect(someObject,"value_changed",anotherObject,"value");

to automatically send value changes from one object to another (see test
for code example). It is quite lightweight (especially it adds no overhead
when you don't use it), and compatible to aRts asynchronous streams (you
can connect an attribute to an async stream and the other way round), works
for any aRts type, and is of course network transparent.

svn path=/trunk/kdelibs/; revision=71765