Fixed "crash when closing the SSL dialog _after_ closing the window".
Whose brilliant idea was kio_uiserver ? :-) On the whole, it created more
trouble than gains imho (although a dialog box shown by the app would
actually block the whole app).

The fix involves setting a pointer to 0 in the destructor and checking
it in the method that gets the result back from the uiserver... A little
bit dangerous, I have to admit, but it works. A better fix would be for
the uiserver to receive the DCOP signal while the msgbox is up, and
delete the msgbox... patches welcome :)

svn path=/trunk/kdelibs/; revision=71864