KToolBarPopupAction (sorry about the name :) was severely missing.
It's a normal action in a menu, but a button+popup in a toolbar.
KonqHistoryAction was implementing that, Konsole has something similar,
and I'll need it for kedittoolbars and later koffice (undo/redo actions).

Werner: don't work on gcommand too much, I'll put kcommand in kdelibs
once I'm done (I hate code duplication) :-)

svn path=/trunk/kdelibs/; revision=72156