Pretty major change in the way that fonts are found.  Previously, all fonts using KFontDialog or KFontChooser were found using XListFonts. Unfortunately, this breaks whenever such things like the RENDER extension are used or KDE runs in (theoretically!) Qt/Embedded or even Windows.
The Right Way(tm) to do this is to just use Qt for all font
information.  Since there is the QFontDatabase class that is *similar*
in functionality to XListFonts, it's clearly the way to go.

For 99.9% of the cases, this does work great -- it's fast and accurate
and works with Xft.  The case where it fails is with fixed width
fonts -- QFontDatabase doesn't contain that knowledge.  So I had to
work around this with a huge hack involving iterating through all
fonts and instantiating a QFontInfo on each.  If you would guess that
this is *very* slow, you would be right.  On the one plus side, I only
found one place in all of KDE that actually uses this feature
(khexedit).

I left the XListFont stuff in for backwards compatability (for now).
Since nothing in KDE uses it, we'll have to get rid of it somehow.

svn path=/trunk/kdelibs/; revision=74508