KHTML will (again) respect the "no underline" setting and we get back our cute little "hover" feature as a nice bonus.
Actually, it was the hover feature that was the problem.  See here:

-        str += "a[href]:hover { text-decoration: underline; }\n";
+        str += "a:hover { text-decoration: underline; }\n";

For some reason that I don't know, if you set *any* properties of
'a[href]', then it overrides all other properties of 'a'.  This meant
that in our attempt to make the "hover" underlined, we forced it to
always be underlined.  I imagine this is a bug somewhere else in the
CSS code... but in any event, this fix works now.

This closes some random bug reports.. I'll track 'em down.

svn path=/trunk/kdelibs/; revision=75424