Initial revision of KPAC, Proxy Auto Configuration for KDE. No GUI and kio_http doesn't use it yet.
KProtocolManager::proxyForURL(const KURL &) uses the
proxy config script if one exists or returns profyFor(url.protocol())
if PAC is not configured. Add Proxy Config Script=URL to config script
to the proxy settings in kioslaverc to enable PAC.

The KJS bindings are in libkpac which is dlopened on demand by
KProtocolManager to avoid kio being dependant on kjs.

svn path=/trunk/kdelibs/; revision=75500