It is now possible to have links Always underlined, Never underlined, or Hover underlined (only underlined when link has focus). Previously, it was only possible to have Always or Hover...
I had to add a private member variable for this implementation.  It
occurs to me, though, that this *might* be a BIC.. but I'm not sure.
The confusion lies in wether or not this file is part of the public
API.  Is it?  If so, why doesn't it have a private d pointer?

svn path=/trunk/kdelibs/; revision=76989