Add: virtual void get_usercredentials()
This is for future methods like kerberos which may want to know various
things that aren't stored in a URL.. like where's my ticket, etc.  Or for
external methods of auth.. or using smart cards, etc.  Hell it could even
be as simple as asking for a user and pass which are absent from the
URL... depends on the module.

svn path=/trunk/kdelibs/; revision=77236