Never *ever* look at MS Frontpage generated HTML... It's just awful. Well, here's a sniplet anyway:
They use <script="font+3"> inside their HTML. khtml recognised
this as the beginning of a script tag, and as such, the following text
never rendered. This patch does behave a bit more as ns does. The '=' sign
is still recognised as part of the tagname, and then we fortunately
don't get a valid tag. If you add a space between script and '=' netscape
fails too, so we are behaving like them now. Anybody speak up if he
sees problems with this patch.

Fixes #18006 btw...

svn path=/trunk/kdelibs/; revision=77531