Yay I can finally commit (ehm, SBC/PacBell/whatever sucks).
Hmm.. finally got xemacs installed.

Oh yeah, time for a KDE related rant:

KWrite.  Was thinking maybe it would help to mmap files when loading them
so as to not require something like 2x the memory of the file to load it.
Well, there's many problems with this theory.

* For whatever reason QIODevice doesn't derive from QObject.  Yay for
that.  What does that mean?  No Qt mock-rtti stuff accessable.  What for?
Well KWrite uses QIODevices all over the place, and for 90% of the time
those qiods are indeed QFiles.  Sure I could use a dynamic cast, but that
would mean real rtti.  Ew.  So I hacked in a few member functions since to
mmap you need the fd (and well calling stat isn't such a bad idea).

* QByteArray deletes its buffer when you change it.  And there's no way to
get around it.  Well when you pass it an mmap'd buffer, all's well.  But
when you want to change buffers... you can't delete something on the
stack.. or something mmapd (is that even considered on the stack?).
Whatever, it's rediculous this can't be changed in QGArray.  This means no
QBuffer, no QByteArray, etc.  Maybe hackery with QTextCodec will work.
Why so inflexible?

And as we wind on down the road...there walks a lady we all know.

svn path=/trunk/kdelibs/; revision=80098