More of this stuff.  It looks nicer *to me*.
"I want a love that's right, but right is only half of what's wrong....
baby I'M in love with you." - the mop top guys

Baby I'm in love with YOU?
BABY I'm in love with you?
Baby I'm in love WITH you?
Baby I'm in LOVE with you?

Boggles the mind doesn't it?

svn path=/trunk/kdelibs/; revision=81039