Here's SOCKS support for KDE.  This is a generic utility which reimplements a bunch of LIBC calls.  It can auto-detect the version of SOCKS installed and will have a configuration panel in the near future.  Right now it works with Dante and NEC SOCKS.  Note that it does not in any way link to these at compile time.  You don't need headers or libraries installed.  Everything is done at runtime.
Please:  anyone on non-linux/glibc2 platforms could you check this code and
see what needs to be done to make it compile for you?  I am sure that there
will be subtle libc and #include differences that need to be addressed.

As of right now, this code is not in the build.  Once we have it compiling
in at least Linux, FreeBSD and Solaris I will commit it.  Then it needs to
be integrated into KSocket, TCPSlaveBase, and any other apps that wish to
use it.

Using it is _very_ easy.  For 99% of the cases, you will just do this:
Look for any calls in ksocks.h that you use straight from libc.  Replace
you calls with KSocks::self()->some_libc_call();

ex:   select()  would become   KSocks::self()->select();  for any cases
where you want SOCKS to work.


All in all, this is a rather simple solution to implementing SOCKS and it
gets rid of the need to have users do LD_PRELOAD hacks.  It's not going to
be 100% clean at first because there will be KSocks references everywhere.
I think it's time to start using KSocket or some other abstracted code for
sockets anyways.


Oh one last thing, I have the modifications to KSocket done but I have not
committed them because I know that Thiago was doing a lot of work here.  The
changes are very minor and can easily be committed later.

svn path=/trunk/kdelibs/; revision=87122