Eewwwwwww. Let's make this compile.
Basically it kdelibs is broken where you don't have getaddrinfo.  That's
not my doing, nor my problem to fix.  However, HAVE_BROKEN_GETADDRINFO
means that we do have getaddrinfo and all the structures, and
kde_getaddrinfo is designed to work around this, but was only defined if
you already had HAVE_GETADDRINFO defined.  Meaning where it would be most
useful it wasn't there, but then again none of the kextsock code calls
getaddrinfo, it calls kde_getaddrinfo which might not actually be there.
Ew.

svn path=/trunk/kdelibs/; revision=90687