While the /tmp files for xauth and iceauth creation were created safely, there were world readable.
Since these files contain the keys to the X and ICE connections, an
attacker can read those files and then snoop the x connection (or even
insert commands).

This patch changes the calls to xauth and iceauth to use popen()
instead of system(".. source tmpfilename"), and also sets the umask
before mkstemp() so files are created mode 0600.

I tested the xauth part, but did not find a ice/dcop example.

svn path=/trunk/kdelibs/; revision=93594