fixed top, parent and self properties which broke when fixing the window object last week :(
Instead of Window objects we store pointers to the KJScript instances in
the KHTMLPart map now. From that we can access the Global object (window)
and its prototype (KJS::Window).

The endless reloads that have appeared over the weekend are gone now. But
I noticed an unexplicable and ugly issue with window.open(): it seems to
work only once :(

svn path=/trunk/kdelibs/; revision=98025