This is wrong but closer to being right than before.
For some reason the KInetSocketAddress return values are all wrong.  I pass
in an address w.x.y.z and I get back something completely random-seeming.
The same if I pass in a hostname.  Anyhow, there are other issues.  I can't
set the IP/host from within kssl.cc because if it's behind a proxy, it will
get the proxy's address instead of the server's address.

svn path=/trunk/kdelibs/; revision=98494