with these changes and the latest ones in libkjs all window.* related problems I know of should be fixed. If not, let's be disciplined enough and add more test cases in khtmltests/ecma.
Using the new Global::extra() member as a pointer to the respective
KHTMLPart we can live without the [[ScriptURL]] hack for security checks.

As the KJS::Window object is at the end of the lookup chain now we simply
always return true in hasProperty() to emulate the IE behaviour of NOT
throwing an exception on undeclared variables. Finally allowed to remove
Niko's hack from kjs/object.cpp in order to become ECMA compliant again.

svn path=/trunk/kdelibs/; revision=98870