Ok, here we go again. Fixed the stupid while-loop:
@@ -932,14 +931,20 @@
   // the job was really running (David, Michael)
   if ( !m_bComplete )
   {
+    KURL::List dirsToBeUpdated;
+
     it = dirs.begin();
     for ( ; it != dirs.end() ; ++it )
     {
       QMap< KIO::ListJob *, QValueList<KIO::UDSEntry> >::Iterator job = d->jobs.begin();
-      while ( job != d->jobs.end() )
-        if ( job.key()->url() == (*it) )
-          updateDirectory( (*it ) );
+      for ( ; job != d->jobs.end(); ++job )
+        if ( job.key()->url().cmp( (*it), true ) )
+          dirsToBeUpdated.append( (*it) );
     }
+
+    it = dirsToBeUpdated.begin();
+    for ( ; it != dirsToBeUpdated.end() ; ++it )
+      updateDirectory( (*it ) );
   }
 }

VERY important is to *not* call updateDirectory immediately because that would
change d->jobs and thus invalidates the iterator - that's why the broken
while-loop worked as well I guess :}

Now always used KURL::cmp insted of KURL::operator==.

Finally added a similar test to FileRenamed() and fixed some indents.

But there's still a bug somewhere - m_lstFileItems does not contain the links
on the desktop! Very strange, because they are listed. Hmm, gave up on
this one for now...

svn path=/trunk/kdelibs/; revision=100750