Support the use of properties like onClick and onMouseOver on elements, to handle events. The property is a javascript function object, so you can do things like:
input1.onClick = function() { alert('test'); }

Also, assigning an on* attribute to an element creates a javascript function
with the body given in the attribute value, and a single parameter, event.
This means that you can now reference the event object from within an event
handler, as many DHTML scripts use.

svn path=/trunk/kdelibs/; revision=104927