For the first time David's analysis of a problem was wrong ;-) I just got a mail from AleXXX with bugreport #29530 (as I found out by comparing the time...) with the following reason:
If selectCurrentItemAndEnableSelectedBySimpleMoveMode() was called (which
only the textview does) KListView::focusOutEvent deselects the current
item, regardless of how many items are selected.
KListView::focusInEvent does nearly the same, it just selects the
current item. So if you have now the sidebar enabled and select a
directory to display in the textview the textview has no focus yet.
Clicking on an item in the textview, holding down the control button, leds to
a focusInEvent selecting the current item (the first one) and the LMB+Ctrl then
selects the second one.

The problem: if you scroll down first, there is no focusInEvent,
so you won't realize the first item gets selected.
Clicking now on an item while pressing Ctrl... well, see #29530 ;)

Fixed with this commit. Waldo, if possible, please move the tag. I think
it's really important for the textview!

svn path=/trunk/kdelibs/; revision=108722