DIGEST-MD5 authentication.
After finding the bug in RFC 2831 and cracking the algorithm it even worked :-)
We follow now the example rather then the description, because at least the
authors of cyrus-sasl also did that. Probably because the all used the buggy
code from 2617 and didn't look at the errata.
No wonder, when looking what company was involved in that RFC.

svn path=/trunk/kdelibs/; revision=110896