Don't use all_libraries multiple times in LDFLAGS.  This gets rid of those -lpthread^1000 in the later libraries (remove all .la files, and remake).
This leaves only one problem: after Bradleys changes -lpthread is in
KDE_MT_LDFLAGS, which leads to -lpthread being in KDE_PLUGIN _and_
all_libraries.  This leads to multiple -lpthread for things which use
$(KDE_PLUGIN) and $(all_libraries), which later again multiplies for anything
linking against those plugins (which anyway should be nothing).  Hmm, either
we leave this as it is, or we need to introduce two KDE_PLUGINS, one for plugins
not using all_libraries, and one (leaving out KDE_MT_LDFLAGS) for those which
do.

svn path=/trunk/kdelibs/; revision=115827