Don't default to 2-pixel wide borders around <img border="0" ...>
Actually I wonder why the default border for images is 2 pixels in the first
place. It horribly breaks layouts that assume there is no border around
images that are not hyperlinks (as is the case since Netscape 1.0 or so
for every browser out there).

There are now still some broken images, but at least it fixes most of them.

@@ -123,7 +123,7 @@ void HTMLImageElementImpl::parseAttribut
         break;
     case ATTR_BORDER:
         // border="noborder" -> border="0"
-        if(attr->value().toInt()) {
+        if(attr->value().toInt() || attr->value() == "0" ) {

svn path=/trunk/kdelibs/; revision=124420