After looking at my diff it occurred to me that the definitive fix might be to use KPushButton everywhere so you always have action to the KGuiItem.
But currently that wouldn't make sense yet: KPushButton only provides this
access by means of its constructor, the KGuiItem is not really embedded as
in the case of KAction. Also, there is no code out there that would make
use of this functionality.

So, in short: the icons-on-buttons issue should be fixed again now, in
a BC manner...

Whether or not to change kdialogbase.h to return KPushButtons instead of
QPushButtons is still open for discussion and might be a better option
for future expansion. For now I postponed this change, though I might
commit it at BIC day :-)

svn path=/trunk/kdelibs/; revision=127277