For more control over which fonts are shown in the font dialog: -  static void getFontList( QStringList &list, bool fixed ); +  static void getFontList( QStringList &list, uint fontListCriteria);
The fonts styles are no longer expected to be the set of 4 normal types:
generic/bold/italic/bold-italic.
Each font is queried which styles are present and those are shown. This
means the user will get what the dialog shows; not a surprise what QT will
select.

If present; a list of fontsizes will be used from the font server; otherwise
the default set will be presented.

svn path=/trunk/kdelibs/; revision=128268