Change how we interpret communication with su to be more robust against unexpected messages.
The previous method expected the first line it received after sending
the password to be "kdesu_stub", otherwise it would assume the password
to be incorrect.

The new method waits till it either gets a line with "kdesu_stub" or
till su exits, in the latter case it assumes the password to be incorrect.

svn path=/trunk/kdelibs/; revision=131278