- Fix BR#15003
<FORM ACTION="mailto:...."> works now.  I implemented "text/plain" which
seems to be common (hence the form_impl addition), and multipart/form-data
which may or may not be correct.  I can't get NS 4.x or Mozilla to work
properly with this for comparison.   Includes security checks to make sure
attach= is not passed, and

GUI: warns the user that the form is being submitted via email

This continues the old behaviour of using kmailservice to send the form.  If
the user wants something different, well, reimplement kmailservice locally :)

Also added code to strip [\n\r\t] and excessive " " from mailto: links when
they are formatted for the statusBar.

svn path=/trunk/kdelibs/; revision=147611