a bunch of small clean ups related to my previous commit as well as things i noticed as i read through the code more.
btw, i can manage to crash it if i toggle the seperate dir view on and off
a bunch of times really quickly. i'm still not sure why exactly, but i'm
looking into it and will eventually track it down. i don't think this is
related to my recent changes, just that before it was harder to toggle
it on/off so quickly (no keyboard shortcut). in any case,
at least i'm getting to know kfile internals quite a bit =)

svn path=/trunk/kdelibs/; revision=164510