This is a relatively minor fix (though annoying to find I assure you!) that was causing completion to not work in certain cases.
i.e. enter '/var/' in the combo box at the top, add 'lo/' -- you are pushed
back to '/var/'.  However, when you add 'l' again (and this depends on
'lo/' not existing in both cases, but 'log/' existing) you are stuck with
'/varl' instead of '/var/l'.  Erm, this fixes that.

+           if(!base.endsWith("/"))
+               base.append('/');
             QString newText = base + complete;

CCMAIL:pupeno@pupeno.com

svn path=/trunk/kdelibs/; revision=165868