Fix for accessing the body of an empty frame/iframe. Testcase: <iframe id="a"></iframe><script>alert(a.document.body);</script>
-        if (m_part->document().isNull()) {
+        if (m_part->document().isNull() || m_part->htmlDocument().body().isNull()) {
           kdDebug(6070) << "Document.write: adding <HTML><BODY> to create document" << endl;

However I don't understand why we don't do this in KHTMLPart::processObjectRequest,
in the about:blank case (under p->m_url = url). Hmm maybe because at the time we
used empty url instead of about:blank. It works too, shall I do that? (it's a hack
to do it in the JS bindings...)

svn path=/trunk/kdelibs/; revision=169481