CCMAIL: tackat@tackat.com In brief: Makes the current icon theme svgz icons have priority to be chosen over the png icons of inherited icon themes (but still, png icons of the current icon theme have preference over svgz icons of the current icon theme)
Long explanation: Currently, when the icon loader searches for an icon, first
it tries to load it in png format, first in the current icon theme, and then
on inherited ones. If it cannot be found in png format, it looks in svgz format,
first in the current icon theme, then in the others. then it looks in svg format
and then in xpm.

This patch fixes that by making it search for the icon first on the current icon
theme in png format, then svgz, svg and xpm, if it cannot be found, then it
looks on the next icon theme, first in png, svgz, svg and finally xpm,
and so on.

I (and others) find it silly that it prefers an icon on an inherited icon theme
just because of the format it's been saved in.

Also, I've noticed that even if the icon loader says it support multiple
inheritance (an icon theme which inherits from more than one icon theme), we've
never supported that, so I clearly don't support it in this code (instead of
giving fake support as the rest of the icon loader does)

svn path=/trunk/kdelibs/; revision=178373