fix for #49701 (nba.com)
The problem was caused by incorrect colspan specifications on the page.
This caused khtml to layout the colspan cell incorrectly.

We now use (RenderTableCell *)-1 in the cell array to indicate that this
position is occupied by a span cell, making it possible to fix this bug.

Did lots of testing, so I hope this won't break anything.

svn path=/trunk/kdelibs/; revision=185993