adds 6 new methods for squeezing strings: right, left and centre versions of squeezing based on font-specific character width (like ems) and on pixel widths. while slower than their character count based counterparts of [lcr]squeeze, they produce much nicer results for lists of items such as in kicker's quickbrowser menus, or klipper's list of items in the clipboard
i've worked fairly hard at making sure they aren't too slow, though. i
average 2200 squeezes of a 1500 character string on my PII-400. with a
random list of variable length strings, these methods usually give =<7
pixels max variation between the resulting squozen strings..

i've gone through every instance of [lcr]squeeze in kdelibs/base and where
it made sense switched to using [lcr]EmSqueeze. i'll be committing these
in the coming weeks after passing them by the maintainers of the respective
code.

small rant: am i the only one who finds it annoying when some new API is
committed to kdelibs and then immediately used in other modules, causing
you to stop your compile if you happen to cvs up in that module and rebuild
that appropriate parts in kdelibs? bah... not that it can always be helped,
of course....

anyways, i'll try not to do that to everyone and give it a week or so
before i start committing changes that use the new API. this'll also give
others time to bitch and complain about the new methods if they feel the
need to ;-P

svn path=/trunk/kdelibs/; revision=198912