Generic preferences dialog stuff... Disclaimer: This is not for actual use yet (except KView, my testing ground ;-) ).
What this is supposed/going to be:
Create a preferences dialog in your app, you can then register pages for
the dialog in your app, part or plugins. The goal is to make merging of
configuration options easier when using plugin/KParts based programs. The
whole thing comes with a plugin selector widget quite similar to the one
that can be found in noatun, and I hope generic enough so that every
current KDE app that has a dialog like this may use it.

The preferences pages of the plugins are handled seperatly like discussed
(very long ago) on core-devel.

I'm committing because I need revision control and KView needs it...

svn path=/trunk/kdelibs/; revision=206061