add more debug, and catch one more case for caching.
Note that you might find the context menu ssl info action is not quite working
in the sense that it displays the status but not the details.  We need to
retrieve this from the same place that we retrieved the session ID from if
and only if the cached session ID is used.

svn path=/trunk/kdelibs/; revision=207415