Merged in the garbage collector from Safari, with some fixes.
The use of the zeroIfFree flag in the union in CollectorCell was
conflicting with ValueImp::refcount. This was causing crashes/leaks in
some situations. Instead of a union we now use the VI_DESTRUCTED flag in
ValueImp, and the _vd pointer to store the link to the next cell in the
free list.

svn path=/trunk/kdelibs/; revision=210661