Another big commit... sorry, i wanted to test it, and as a consequence it just kept getting bigger :(
The good news is that this _dramatically_ improves the speed of the syntax
highlighter (in one extreme test case, down from many minutes to 3 seconds),
leading to very fast load times (if you can't see it, chances
are you haven't used kate before ;)

The problem was elucidated with valgrind and kcachegrind, many thanks to the
authors of those programs.  Basically it was spending (in the worst case
scenario) 99.7% of the time in memcpy(), because the textline is stored as
a QMemArray<QChar>, and each regexp match needed the string copied (this
was done once per char in the context, for each line).  For now, I've not
changed the TextLine storage, but only made it create the QString once per
line during parsing.  If/when the textline gets switched over there should
be more, but less noticeable gains.

Also in this commit is a change to KateTextCursor, makingn line and col
protected and provinding retrieve / set functions... this is part of the
arbitrary highlighting interface development.  It's a big diff but not much
changes really.

Other small things: made undo merging not occur after 5 seconds of no typing,
more work on the arbitrary highlighting interface, a whatsthis documentation,
and probably some other small things I forgot.

Enjoy your free speed bost, and email me or the list with any regressions:

CCMAIL:kwrite-devel@mail.kde.org

svn path=/trunk/kdelibs/; revision=213014