Make the positioning of the completion popup in comboboxes match that of the combobox dropdowns.
Remove dead code in eventFilter.

BTW, looks like the last commit introduced a small bug -- when the completion box is shown
above the combobox (when the combo is at the bottom of the screen) -- typing shrinks the completion list
making it "fly" high above the combo. [A similar problem exists regardless of that change when the completion box expands,
but that's far harder to trigger].

svn path=/trunk/kdelibs/; revision=213654