Fix Bug 55571: https authentication through proxy fails
(This patch basically reverts r1.546 and r1.544)
m_bNeedTunnel was incorrectly used to check whether we were busy connecting
to a proxy or sending the actual request. Moved some code around so that it
only gets executed when sending the actual request so that it isn't necassery
any more to abuse m_bNeedTunnel for this.

The resetting of m_bNeedTunnel caused bug 55571, it isn't necassery to reset
it, the resetting was a poor attempt to let m_bNeedTunnel serve as a status
indicator for something it wasn't intended to indicate.

svn path=/trunk/kdelibs/; revision=214519