remove _KONQBLOCK and handle stray table contents implicitly.
I've tried to understand the complicated code in the Safari tree for several
hours and then got a big headache. The discardedAttrs handling doesn't look
like it would make any difference, and the inStrayTableContent flag does
not need to be stored on the stack, we will rediscover the malformed table
structure everytime it matters anyway.

So this is my attempt, and it should do it as well. Maybe not. We'll see.
CCMAIL: khtml-devel@kde.org

svn path=/trunk/kdelibs/; revision=220897