Ok, no response yet, so I'll do it. From my comment to #46477:
Now, I don't quite understand the bugreport, but judging from the backtrace
what happened is this: directory u was opened, the job started listing the
items and just before it sent "." to KDirLister (which causes rootFileItem to
be set) a fam event came in telling it to delete u. The job was stopped and
KDirListerCache made the KDirLister emit deleteItem(rootFileItem) which is not
set yet.

So yes, I guess we need a check for d->rootFileItem being non-NULL.

svn path=/trunk/kdelibs/; revision=228439