A patch from Stanislav Kljuhhin to make this compile with --enable-debug.
You know as <cough>cool</cough> as this whole d() macro thing instead of
d-pointers is, well, it's (a) hard to read and (b) it seems that most of
these classes (which are in installed headers) aren't using it.  This
looks like it's asking for some BIC badness...</rant>

As it currently stands this define b0rks every variable named "d" that
gets tossed into the same compilation unit.

CCMAIL:crz@starman.ee, lypanov@kde.org

svn path=/trunk/kdelibs/; revision=229398