TT says that they won't fix this one before 3.3, so I'm (sadly) doing the fix here.  The affects at least 5 apps in CVS.
Right now if a MainWindow is hidden that has floating toolbars those are
hidden, but aren't restored when it is shown.  In some apps this requires
editing the config file to get the toolbar back.

See the mentioned bug for the qt-bugs mails.  Also did some s/TRUE/true/
s/FALSE/false/ while I was in there.

CCMAIL:60645-done@bugs.kde.org

svn path=/trunk/kdelibs/; revision=239217