:first-letter is finally supported. It may seem unnecessary, but I want caret mode to eventually support *everything*, so :first-letter is an arbitrary start.
For this to be done, I added a m_firstLetter member to RenderFlow, which is set in RenderFlow::addChildToFlow. Furthermore, I established a one to one relationship between RenderText string offsets and DOM offsets which resulted in the addition of RenderText::(set|is)ForcedMinOffset. This method specifies the first offset after which the string will be used for rendering and layouting. The relevant places in bidi.cpp have been adapted to it.

The former anonymous RenderText containing the first-letter now points to the DOM node. Otherwise, EditableInlineBoxIterator would ignore it, making seamless navigation impossible.

RenderText::findInlineTextBox now also considers :first-letter text boxes, if told to do so.
So does CharacterDataImpl::minOffset.

... and it breaks vertical navigation ... it seems to be impossible that *nothing* breaks on a change.

svn path=/trunk/kdelibs/; revision=247953