Finally implement the error notification portion of the statusbar extension. It's not pretty, but it is basically functional.  I will add the ability to disable it soon.
If someone is real keen (pmk? :)), it would be nice to have a way to jump to
the error in the debugger from here.  Also I think this eliminates the need for
the error window in ecma/.  This is a nicer approach IMHO (less intrusive).

Is there a better way to display these error messages?  (perhaps more detailed?)

CCMAIL: 64683-done@bugs.kde.org, pmk@post.com

svn path=/trunk/kdelibs/; revision=255284