Remove the extra columns, setFullWidth on the listview's remaining colum, make the layout vertical and remove the splitter, which is now pretty useless.
Matthias said Kopete is the only app using this class with embedded KCMs,
so the changed look isn't a problem. On the contrary, it's a discussion and
a quick hack in Kopete that spawned these changes in the first place.

Now I only need to split off Kopete's plugin config in a separate 'Configure
Plugins...' dialog.

svn path=/trunk/kdelibs/; revision=259734