OK, here's the real fix:
I changed the parent of the treelist. Too bad that there was code
relying on a fixed QObject tree structure. Restore the old behaviour and
fix another crash by keeping a pointer to the splitter and using that
instead of relying on the parent of the parent being the right widget.

CCMAIL: 66431@bugs.kde.org

svn path=/trunk/kdelibs/; revision=265511