This is the more enhanced version of Japanese encoding detection code.
The original one cannot detect UTF-8 correctly, so I improved this and now,
KHTML can detect it.

This routine(guess_jp) is derived from the one which is used in Gauche, the
Scheme interpreter, BSD licensed(so I include the copyright).
I contacted Takumi Asaki(asaki@kde.gr.jp), who is the author of original code,
and he said OK to commit this code.

svn path=/trunk/kdelibs/; revision=284763