Removed the old perl script.
Added a new C++ version of the script that is mostly the same but with
the noted changes:
	* The input file is a correctly formed KDE ini style file
	* The binary makekdewidgets is installed to the system for
	  general use.
Added tests that ensure the output is correct for building with
the API TrollTech has defined for building widget plugins.

Updated the kde.widgets file to reflect the new format.  Basicly
the first 5 lines where modified slightly.

Updated the README file to reflect the changed options.  Functionaly
its operation is identical to the old perl version, but it has an
option to embed pixmaps.  This is handy if someone wishes to write
a GUI plugin that will work with Qt only.

Some things to note.  People may have problems building this stuff if
they have stale files laying arround.  This has been tested with
unsermake, without unsermake and both builddir != srcdir and builddir
== srcdir.  Chances are if there is a problem with this code its your
local system.  So before you "fix" any thing please forward the patch
to me to ensure your not breaking things for everyone else.

This program will be for general use, and I will have a complete howto
posted shortly.  Ideally we can now add widget plugins for KOffice, and
kde-edu next ;)

CCMAIL: kde-core-devel@kde.org

svn path=/trunk/kdelibs/; revision=285977