Added a new function KIconViewItem::setPixmapSize( const QSize& ).
You can now specify a custom size for the pixmap. The geometry of the
item will be calculated to let a pixmap of the given size fit in the
iconView without needing an update.
This may be useful if you want to change the pixmap later without
breaking the layout. A possible use of this function is to replace a
fileItem icon by a larger pixmap (preview).

This should be useful for Konqueror when using previews with boosted size.

CCMAIL: faure@kde.org

svn path=/trunk/kdelibs/; revision=286829