Fix test suite.
Tobias, the vcard5.vcf test fails because one line of the binary encoded
data is written differently than by the old parser. Could you please have
a look if that is a bug of the old or your new parser, fix the new parser
if required, and readd the test to the test suite?

CCMAIL:tokoe@kde.org

svn path=/trunk/kdelibs/; revision=292459